// types/sunoUserStatsTypes.ts
export interface TagStat {
  name: string;
  count: number;
  totalUpvotes: number;
  totalPlays: number; 
  totalComments: number; 
  avgPlays: number;
  avgUpvotes: number;
  avgComments: number;
  avgUpvoteRate?: number; // New: (Total Upvotes for Tag / Total Plays for Tag) * 100
  avgCommentRate?: number; // New: (Total Comments for Tag / Total Plays for Tag) * 100
}